Convert ProfileSwitches

+++ database/src/main/java/info/nightscout/androidaps/database/embedments/InsulinConfiguration.kt
@@ -2,6 +2,6 @@ package info.nightscout.androidaps.database.embedments
 
 data class InsulinConfiguration(
         var insulinLabel: String,
-        var insulinEndTime: Int,
-        var peak: Int
+        var insulinEndTime: Long,
+        var peak: Long
 )
